Split Containerfile class
- split out Containerfile class into separate file
- move logic for writing container file steps from AnsibleBuilder to Containerfile
